task 6

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ long int y;
 long long int z;
 float a;
 char b;
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(x));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(y));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(z));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of a char: %d byte(s)\n", sizeof(b));
+printf("Size of a int: %d byte(s)\n", sizeof(x));
+printf("Size of a long int: %d byte(s)\n", sizeof(y));
+printf("Size of a long long int: %d byte(s)\n", sizeof(z));
+printf("Size of a float: %d byte(s)\n", sizeof(a));
 return (0);
 }
